metadata processing centralised; etc
